experiencemaker/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ExperienceMaker"
version = "0.1.1"
description = "make experience from trajectory"
authors = [{ name = "experiencemaker team", email = "experiencemaker@alibaba-inc.com" }]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.12"

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]

dependencies = [
    "dashscope>=1.19.1",
    "elasticsearch>=8.14.0",
    "fastapi>=0.115.13",
    "fastmcp>=2.10.6",
    "loguru>=0.7.3",
    "mcp>=1.9.4",
    "numpy>=2.3.0",
    "openai>=1.88.0",
    "pydantic>=2.11.7",
    "PyYAML>=6.0.2",
    "Requests>=2.32.4",
    "uvicorn>=0.34.3",
    "setuptools>=75.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["experiencemaker*"]
exclude = ["cookbook*"]

[tool.setuptools.package-data]
experiencemaker = [
    "config/*.yaml", 
    "op/**/*.yaml",
]

[project.scripts]
experiencemaker = "experiencemaker.app:main"
experiencemaker_mcp = "experiencemaker.mcp_server:main"
